+++ src/pybind/feat/feature_pybind.h
@@ -1,0 +1,25 @@
+// pybind/feat/feature_pybind.h
+
+// Copyright 2019   Microsoft Corporation (author: Xingyu Na)
+
+// See ../../../COPYING for clarification regarding multiple authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+// THIS CODE IS PROVIDED *AS IS* BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, EITHER EXPRESS OR IMPLIED, INCLUDING WITHOUT LIMITATION ANY IMPLIED
+// WARRANTIES OR CONDITIONS OF TITLE, FITNESS FOR A PARTICULAR PURPOSE,
+// MERCHANTABLITY OR NON-INFRINGEMENT.
+// See the Apache 2 License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef KALDI_PYBIND_FEAT_FEATURE_PYBIND_H_
+#define KALDI_PYBIND_FEAT_FEATURE_PYBIND_H_
+
+#include "pybind/kaldi_pybind.h"
+
+void pybind_feature(py::module& m);
+
+#endif  // KALDI_PYBIND_FEAT_FEATURE_PYBIND_H_
