+++ src/pybind/feat/wave_reader_pybind.cc
@@ -16,8 +16,9 @@
 // limitations under the License.
 
 #include "feat/wave_reader_pybind.h"
+
 #include "feat/wave-reader.h"
-#include "util/table-types.h"
+#include "util/kaldi_table_pybind.h"
 
 using namespace kaldi;
 
@@ -55,51 +56,9 @@ void pybind_wave_reader(py::module& m) {
       .def("CopyFrom", &WaveData::CopyFrom)
       .def("Swap", &WaveData::Swap);
 
-  py::class_<SequentialTableReader<WaveHolder>>(m, "SequentialWaveReader")
-      .def(py::init<>())
-      .def(py::init<const std::string&>(), py::arg("rspecifier"))
-      .def("Open", &SequentialTableReader<WaveHolder>::Open, py::arg("rspecifier"))
-      .def("Done", &SequentialTableReader<WaveHolder>::Done)
-      .def("Key", &SequentialTableReader<WaveHolder>::Key)
-      .def("FreeCurrent", &SequentialTableReader<WaveHolder>::FreeCurrent)
-      .def("Value", &SequentialTableReader<WaveHolder>::Value,
-           py::return_value_policy::reference)
-      .def("Next", &SequentialTableReader<WaveHolder>::Next)
-      .def("IsOpen", &SequentialTableReader<WaveHolder>::IsOpen)
-      .def("Close", &SequentialTableReader<WaveHolder>::Close);
-
-  py::class_<RandomAccessTableReader<WaveHolder>>(m, "RandomAccessWaveReader")
-      .def(py::init<>())
-      .def(py::init<const std::string&>(), py::arg("rspecifier"))
-      .def("Open", &RandomAccessTableReader<WaveHolder>::Open, py::arg("rspecifier"))
-      .def("IsOpen", &RandomAccessTableReader<WaveHolder>::IsOpen)
-      .def("Close", &RandomAccessTableReader<WaveHolder>::Close)
-      .def("HasKey", &RandomAccessTableReader<WaveHolder>::HasKey, py::arg("key"))
-      .def("Value", &RandomAccessTableReader<WaveHolder>::Value,
-           py::return_value_policy::reference);
-
-  py::class_<SequentialTableReader<WaveInfoHolder>>(m, "SequentialWaveInfoReader")
-      .def(py::init<>())
-      .def(py::init<const std::string&>(), py::arg("rspecifier"))
-      .def("Open", &SequentialTableReader<WaveInfoHolder>::Open, py::arg("rspecifier"))
-      .def("Done", &SequentialTableReader<WaveInfoHolder>::Done)
-      .def("Key", &SequentialTableReader<WaveInfoHolder>::Key)
-      .def("FreeCurrent", &SequentialTableReader<WaveInfoHolder>::FreeCurrent)
-      .def("Value", &SequentialTableReader<WaveInfoHolder>::Value,
-           py::return_value_policy::reference)
-      .def("Next", &SequentialTableReader<WaveInfoHolder>::Next)
-      .def("IsOpen", &SequentialTableReader<WaveInfoHolder>::IsOpen)
-      .def("Close", &SequentialTableReader<WaveInfoHolder>::Close);
-
-  py::class_<RandomAccessTableReader<WaveInfoHolder>>(m, "RandomAccessWaveInfoReader")
-      .def(py::init<>())
-      .def(py::init<const std::string&>(), py::arg("rspecifier"))
-      .def("Open", &RandomAccessTableReader<WaveInfoHolder>::Open, py::arg("rspecifier"))
-      .def("IsOpen", &RandomAccessTableReader<WaveInfoHolder>::IsOpen)
-      .def("Close", &RandomAccessTableReader<WaveInfoHolder>::Close)
-      .def("HasKey", &RandomAccessTableReader<WaveInfoHolder>::HasKey, py::arg("key"))
-      .def("Value", &RandomAccessTableReader<WaveInfoHolder>::Value,
-           py::return_value_policy::reference);
-
+  pybind_sequential_table_reader<WaveHolder>(m, "_SequentialWaveReader");
+  pybind_sequential_table_reader<WaveInfoHolder>(m, "_SequentialWaveInfoReader");
+  pybind_random_access_table_reader<WaveHolder>(m, "_RandomAccessWaveReader");
+  pybind_random_access_table_reader<WaveInfoHolder>(m, "_RandomAccessWaveInfoReader");
 }
